Move bits into graphics and routing
