+++ trview.app/Geometry/TransparencyBuffer.cpp
@@ -1,7 +1,7 @@
 #include "TransparencyBuffer.h"
 
 #include <trview.app/Camera/ICamera.h>
-#include <trview.app/ILevelTextureStorage.h>
+#include <trview.app/Graphics/ILevelTextureStorage.h>
 #include <trview.app/Geometry/MeshVertex.h>
 
 #include <algorithm>
