+++ trview.app/Graphics/ILevelTextureStorage.cpp
