+++ trview.app/Graphics/ILevelTextureStorage.h
