+++ trview.app/Graphics/ITextureStorage.cpp
