+++ trview.app/Graphics/ITextureStorage.h
