+++ trview.app/Graphics/SectorHighlight.cpp
