+++ trview.app/Graphics/SectorHighlight.h
