+++ trview.app/Graphics/SelectionRenderer.cpp
