+++ trview.app/Graphics/SelectionRenderer.h
