+++ trview.app/Routing/Waypoint.cpp
