+++ trview.app/Routing/Waypoint.h
