+++ trview.app/trview.app.vcxproj
@@ -34,16 +34,16 @@
     <ClCompile Include="Geometry\PickResult.cpp" />
     <ClCompile Include="Geometry\TransparencyBuffer.cpp" />
     <ClCompile Include="Geometry\TransparentTriangle.cpp" />
-    <ClCompile Include="ILevelTextureStorage.cpp" />
-    <ClCompile Include="ITextureStorage.cpp" />
+    <ClCompile Include="Graphics\ILevelTextureStorage.cpp" />
+    <ClCompile Include="Graphics\ITextureStorage.cpp" />
+    <ClCompile Include="Graphics\SectorHighlight.cpp" />
+    <ClCompile Include="Graphics\SelectionRenderer.cpp" />
     <ClCompile Include="Menus\DirectoryListing.cpp" />
     <ClCompile Include="Menus\FileDropper.cpp" />
     <ClCompile Include="Menus\LevelSwitcher.cpp" />
     <ClCompile Include="Menus\RecentFiles.cpp" />
     <ClCompile Include="Menus\ViewMenu.cpp" />
     <ClCompile Include="Route.cpp" />
-    <ClCompile Include="SectorHighlight.cpp" />
-    <ClCompile Include="SelectionRenderer.cpp" />
     <ClCompile Include="Tools\Compass.cpp" />
     <ClCompile Include="Tools\Measure.cpp" />
     <ClCompile Include="Tools\Toolbar.cpp" />
@@ -89,16 +89,16 @@
     <ClInclude Include="Geometry\TransparencyBuffer.h" />
     <ClInclude Include="Geometry\TransparentTriangle.h" />
     <ClInclude Include="Geometry\Triangle.h" />
-    <ClInclude Include="ILevelTextureStorage.h" />
-    <ClInclude Include="ITextureStorage.h" />
+    <ClInclude Include="Graphics\ILevelTextureStorage.h" />
+    <ClInclude Include="Graphics\ITextureStorage.h" />
+    <ClInclude Include="Graphics\SectorHighlight.h" />
+    <ClInclude Include="Graphics\SelectionRenderer.h" />
     <ClInclude Include="Menus\DirectoryListing.h" />
     <ClInclude Include="Menus\FileDropper.h" />
     <ClInclude Include="Menus\LevelSwitcher.h" />
     <ClInclude Include="Menus\RecentFiles.h" />
     <ClInclude Include="Menus\ViewMenu.h" />
     <ClInclude Include="Route.h" />
-    <ClInclude Include="SectorHighlight.h" />
-    <ClInclude Include="SelectionRenderer.h" />
     <ClInclude Include="Tools\Compass.h" />
     <ClInclude Include="Tools\Measure.h" />
     <ClInclude Include="Tools\Toolbar.h" />
