+++ trview/Entity.cpp
@@ -4,7 +4,7 @@
 #include <stack>
 #include <algorithm>
 
-#include <trview.app/ILevelTextureStorage.h>
+#include <trview.app/Graphics/ILevelTextureStorage.h>
 #include "IMeshStorage.h"
 #include <trview.app/Camera/ICamera.h>
 #include <trview.app/Geometry/Mesh.h>
