+++ trview/Room.cpp
@@ -4,7 +4,7 @@
 #include "Entity.h"
 #include "Level.h"
 
-#include <trview.app/ILevelTextureStorage.h>
+#include <trview.app/Graphics/ILevelTextureStorage.h>
 #include "IMeshStorage.h"
 #include <trview.app/Camera/ICamera.h>
 #include <trview.app/Geometry/Mesh.h>
