+++ trview/Viewer.h
@@ -37,7 +37,7 @@
 #include <trview.app/Windows/RouteWindowManager.h>
 #include <trview.app/Menus/ViewMenu.h>
 #include <trview.app/Geometry/Picking.h>
-#include <trview.app/SectorHighlight.h>
+#include <trview.app/Graphics/SectorHighlight.h>
 #include <trview.app/UI/ViewerUI.h>
 
 namespace trview
